Use some types other an `int`
Thrown in some unsigned and short types

+++ sycl/test-e2e/Graph/Explicit/sub_graph.cpp
@@ -10,7 +10,7 @@
 int main() {
   queue TestQueue;
 
-  using T = int;
+  using T = short;
 
   // Values used to modify data inside kernels.
   const int ModValue = 7;
